# swagger.yaml
openapi: 3.0.0
info:
  title: Ytdlp wrapper api
  description: api that download youtube video using ytdlp
  version: 1.0.0
servers:
  - url: 'http://localhost:8000'
paths:
  /download:
    get:
      summary: Download YouTube video
      description: Initiates download of a YouTube video with optional presets or parameters
      parameters:
        - name: link
          in: query
          required: true
          description: YouTube video URL
          schema:
            type: string
            example: "https://www.youtube.com/watch?v=dQw4w9WgXcQ"
        - name: preset
          in: query
          required: false
          description: Whether to use default preset configuration
          schema:
            type: boolean
            default: false
        - name: params
          in: query
          required: false
          description: Additional yt-dlp parameters
          schema:
            type: string
            example: "-f 'bestvideo[ext=mp4]+bestaudio[ext=m4a]/mp4'"
      responses:
        '202':
          description: Download initiated
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "Downloading..."
                  message:
                    type: string
                    example: "Got the request, file is now downloading."
                  id:
                    type: number
                    example: 1644841203000
                  ytid:
                    type: string
                    example: "dQw4w9WgXcQ"
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Invalid Request"
  
  /status/{ID}:
    get:
      summary: Check download status
      description: Checks the status of a download using its ID
      parameters:
        - name: ID
          in: path
          required: true
          description: Download ID
          schema:
            type: string
      responses:
        '200':
          description: Success response
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: number
                    example: 1644841203000
                  filename:
                    type: string
                    example: "output-dQw4w9WgXcQ.mp4"
                  status:
                    type: boolean
                    example: true
                  ytid:
                    type: string
                    example: "dQw4w9WgXcQ"
        '400':
          description: Bad request
        '404':
          description: Download not found

  /redownload/{ID}:
    get:
      summary: Re-download a file
      description: Download a file that was previously processed, using its ID
      parameters:
        - name: ID
          in: path
          required: true
          description: Download ID
          schema:
            type: string
      responses:
        '200':
          description: File download response
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '202':
          description: File still downloading
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "File is still downloading, please wait!"
        '400':
          description: Bad request
        '404':
          description: File not found

  /delete/{ID}:
    post:
      summary: Delete a download
      description: Delete a downloaded file and remove it from the database
      parameters:
        - name: ID
          in: path
          required: true
          description: Download ID
          schema:
            type: string
      responses:
        '200':
          description: Success response
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "File Deleted"
        '400':
          description: Bad request
        '404':
          description: File not found